requirements.txt
openai
fastapi
uvicorn
python-multipart
numpy
websockets
scipy
google-generativeai